Use new endpoint for series image upload
